[MAINTENANCE] Move `convert_to_json_serializable` to top-level utils package (#9933)
