+++ assets/scripts/build_package_gallery.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING, List
 import pip
 from great_expectations_contrib.commands import read_package_from_file, sync_package
 
-from great_expectations.core.util import convert_to_json_serializable
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     from great_expectations_contrib.package import (
