+++ contrib/cli/great_expectations_contrib/commands.py
@@ -8,7 +8,7 @@ from typing import NamedTuple
 import click
 from cookiecutter.main import cookiecutter
 
-from great_expectations.core.util import convert_to_json_serializable
+from great_expectations.util import convert_to_json_serializable
 
 from .package import GreatExpectationsContribPackageManifest
