+++ contrib/experimental/great_expectations_experimental/expectations/expect_queried_column_pair_values_to_be_both_filled_or_null.py
@@ -1,6 +1,5 @@
 from typing import Optional, Union
 
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.exceptions.exceptions import (
     InvalidExpectationConfigurationError,
 )
@@ -12,6 +11,7 @@ from great_expectations.expectations.expectation import (
 from great_expectations.expectations.expectation_configuration import (
     ExpectationConfiguration,
 )
+from great_expectations.util import convert_to_json_serializable
 
 
 class ExpectQueriedColumnPairValuesToBeBothFilledOrNull(QueryExpectation):
