+++ contrib/experimental/great_expectations_experimental/expectations/expect_queried_column_values_to_exist_in_second_table_column.py
@@ -1,11 +1,11 @@
 from typing import Union
 
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.execution_engine import ExecutionEngine
 from great_expectations.expectations.expectation import (
     ExpectationValidationResult,
     QueryExpectation,
 )
+from great_expectations.util import convert_to_json_serializable
 
 
 class ExpectQueriedColumnValuesToExistInSecondTableColumn(QueryExpectation):
