+++ contrib/experimental/great_expectations_experimental/expectations/expect_queried_custom_query_to_return_num_rows.py
@@ -1,6 +1,5 @@
 from typing import Optional, Union
 
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.execution_engine import ExecutionEngine
 from great_expectations.expectations.expectation import (
     ExpectationValidationResult,
@@ -10,6 +9,7 @@ from great_expectations.expectations.expectation import (
 from great_expectations.expectations.expectation_configuration import (
     ExpectationConfiguration,
 )
+from great_expectations.util import convert_to_json_serializable
 
 
 class ExpectQueriedCustomQueryToReturnNumRows(QueryExpectation):
