+++ contrib/experimental/great_expectations_experimental/expectations/expect_queried_slowly_changing_table_to_have_no_gaps.py
@@ -1,7 +1,6 @@
 from datetime import datetime, timedelta
 from typing import Optional, Union
 
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.exceptions.exceptions import (
     InvalidExpectationConfigurationError,
 )
@@ -13,6 +12,7 @@ from great_expectations.expectations.expectation import (
 from great_expectations.expectations.expectation_configuration import (
     ExpectationConfiguration,
 )
+from great_expectations.util import convert_to_json_serializable
 
 
 class ExpectQueriedSlowlyChangingTableToHaveNoGaps(QueryExpectation):
