+++ contrib/experimental/great_expectations_experimental/expectations/expect_query_count_with_filter_to_meet_threshold.py
@@ -6,7 +6,6 @@ For detailed instructions on how to use it, please see:
 
 from typing import Optional, Union
 
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.exceptions.exceptions import (
     InvalidExpectationConfigurationError,
 )
@@ -18,6 +17,7 @@ from great_expectations.expectations.expectation import (
 from great_expectations.expectations.expectation_configuration import (
     ExpectationConfiguration,
 )
+from great_expectations.util import convert_to_json_serializable
 
 
 class ExpectQueryCountWithFilterToMeetThreshold(QueryExpectation):
