+++ great_expectations/core/batch.py
@@ -21,10 +21,13 @@ from great_expectations._docs_decorators import deprecated_argument
 from great_expectations.compatibility import pyspark
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.id_dict import BatchKwargs, BatchSpec, IDDict
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.exceptions import InvalidBatchIdError
 from great_expectations.types import DictDot, SerializableDictDot, safe_deep_copy
-from great_expectations.util import deep_filter_properties_iterable, load_class
+from great_expectations.util import (
+    convert_to_json_serializable,
+    deep_filter_properties_iterable,
+    load_class,
+)
 
 if TYPE_CHECKING:
     from typing_extensions import NotRequired, TypeAlias
