+++ great_expectations/core/expectation_validation_result.py
@@ -20,11 +20,7 @@ from great_expectations.core.batch import (  # noqa: TCH001
 )
 from great_expectations.core.id_dict import BatchSpec  # noqa: TCH001
 from great_expectations.core.run_identifier import RunIdentifier  # noqa: TCH001
-from great_expectations.core.util import (
-    convert_to_json_serializable,
-    ensure_json_serializable,
-    in_jupyter_notebook,
-)
+from great_expectations.core.util import in_jupyter_notebook
 from great_expectations.data_context.util import instantiate_class_from_config
 from great_expectations.exceptions import ClassInstantiationError
 from great_expectations.render import (
@@ -35,6 +31,10 @@ from great_expectations.render import (
     RenderedAtomicContentSchema,
 )
 from great_expectations.types import SerializableDictDot
+from great_expectations.util import (
+    convert_to_json_serializable,
+    ensure_json_serializable,
+)
 
 if TYPE_CHECKING:
     from great_expectations.expectations.expectation_configuration import (
