+++ great_expectations/core/serializer.py
@@ -18,7 +18,7 @@ serialized_value = serializer.serialize(config)
 import abc
 from typing import TYPE_CHECKING
 
-from great_expectations.core.util import convert_to_json_serializable
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     from marshmallow import Schema
