+++ great_expectations/core/suite_parameters.py
@@ -26,8 +26,8 @@ from pyparsing import (
     dictOf,
 )
 
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.exceptions import SuiteParameterError
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     from typing_extensions import TypeAlias, TypeGuard
