+++ great_expectations/data_context/types/base.py
@@ -43,9 +43,8 @@ from great_expectations._docs_decorators import public_api
 from great_expectations.compatibility import pyspark
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.configuration import AbstractConfig, AbstractConfigSchema
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.types import DictDot, SerializableDictDot
-from great_expectations.util import deep_filter_properties_iterable
+from great_expectations.util import convert_to_json_serializable, deep_filter_properties_iterable
 
 if TYPE_CHECKING:
     from io import TextIOWrapper
