+++ great_expectations/execution_engine/sparkdf_execution_engine.py
@@ -40,10 +40,7 @@ from great_expectations.core.id_dict import IDDict
 from great_expectations.core.metric_domain_types import (
     MetricDomainTypes,  # noqa: TCH001
 )
-from great_expectations.core.util import (
-    AzureUrl,
-    convert_to_json_serializable,
-)
+from great_expectations.core.util import AzureUrl
 from great_expectations.exceptions import (
     BatchSpecError,
     ExecutionEngineError,
@@ -68,6 +65,7 @@ from great_expectations.expectations.row_conditions import (
     RowConditionParserType,
     parse_condition_to_spark,
 )
+from great_expectations.util import convert_to_json_serializable
 from great_expectations.validator.computed_metric import MetricValue  # noqa: TCH001
 from great_expectations.validator.metric_configuration import (
     MetricConfiguration,  # noqa: TCH001
