+++ great_expectations/expectations/row_conditions.py
@@ -22,8 +22,8 @@ import great_expectations.exceptions as gx_exceptions
 from great_expectations.compatibility.pyspark import functions as F
 from great_expectations.compatibility.sqlalchemy import sqlalchemy as sa
 from great_expectations.compatibility.typing_extensions import override
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.types import SerializableDictDot
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     from great_expectations.compatibility import pyspark, sqlalchemy
