+++ great_expectations/experimental/rule_based_profiler/estimators/numeric_range_estimation_result.py
@@ -4,8 +4,8 @@ from dataclasses import asdict, dataclass
 from typing import TYPE_CHECKING, List, Union
 
 from great_expectations.compatibility.typing_extensions import override
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.types import DictDot
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     import numpy as np
