+++ great_expectations/experimental/rule_based_profiler/helpers/configuration_reconciliation.py
@@ -6,11 +6,12 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 from great_expectations.compatibility.typing_extensions import override
-from great_expectations.core.util import convert_to_json_serializable, nested_update
+from great_expectations.core.util import nested_update
 from great_expectations.experimental.rule_based_profiler.helpers.util import (
     convert_variables_to_dict,
 )
 from great_expectations.types import SerializableDictDot
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     from great_expectations.experimental.rule_based_profiler.parameter_container import (
