+++ great_expectations/experimental/rule_based_profiler/helpers/runtime_environment.py
@@ -6,11 +6,11 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.metric_domain_types import MetricDomainTypes
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.experimental.rule_based_profiler.helpers.util import (
     convert_variables_to_dict,
 )
 from great_expectations.types import SerializableDictDot
+from great_expectations.util import convert_to_json_serializable
 
 if TYPE_CHECKING:
     from great_expectations.experimental.rule_based_profiler.rule import Rule
