+++ great_expectations/experimental/rule_based_profiler/rule/rule.py
@@ -6,7 +6,6 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Union
 
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.util import (
-    convert_to_json_serializable,
     determine_progress_bar_method_by_environment,
 )
 from great_expectations.experimental.rule_based_profiler.config.base import (
@@ -29,6 +28,7 @@ from great_expectations.experimental.rule_based_profiler.parameter_container imp
 from great_expectations.experimental.rule_based_profiler.rule.rule_state import RuleState
 from great_expectations.types import SerializableDictDot
 from great_expectations.util import (
+    convert_to_json_serializable,
     deep_filter_properties_iterable,
     measure_execution_time,
 )
