+++ great_expectations/types/attributes.py
@@ -4,8 +4,8 @@ import logging
 
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core import IDDict
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.types import SerializableDotDict
+from great_expectations.util import convert_to_json_serializable
 
 logger = logging.getLogger(__name__)
