+++ great_expectations/validator/exception_info.py
@@ -4,8 +4,8 @@ import json
 
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core import IDDict
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.types.base import SerializableDotDict
+from great_expectations.util import convert_to_json_serializable
 
 
 class ExceptionInfo(SerializableDotDict):
