+++ tests/core/test_batch_related_objects.py
@@ -11,8 +11,8 @@ from great_expectations.core.batch import (
 )
 from great_expectations.core.batch_spec import RuntimeDataBatchSpec
 from great_expectations.core.id_dict import deep_convert_properties_iterable_to_id_dict
-from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.exceptions import InvalidBatchSpecError
+from great_expectations.util import convert_to_json_serializable
 
 
 @pytest.mark.unit
